Add the ActionForward class, which represents a path to which control should be forwarded, or redirected, by the controller servlet.
Modfiy Action.perform() to return an ActionForward() entry, typically
configured for this <action> in the XML configuration file.

Modify ActionServlet to respect the value returned by the perform()
method of the called Action class.

Modify the example application to use the new technique for registering
any number of logically named <forward> entries, for use in determining
the next stage of the user interface dynamically.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47687 13f79535-47bb-0310-9956-ffa450edef68