Create two subclasses of ActionForward with different default values for the "redirect" property.  Now you can essentially change the global default for whether to use RequestDispatcher.forward() or response.sendRedirect() by changing which class you define as the value of the "forward" initialization parameter to the controller servlet.
Any of the ActionForward implementations let you specifically set the value
of the "redirect" property if you want to ensure that certain cases are
always handled the same way.

The default if you do nothing is that ActionForward objects have their
"redirect" property set to false, so that RequestDispatcher.forward()
will be used.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47702 13f79535-47bb-0310-9956-ffa450edef68