Encapsulate the global "forwards" and "mappings" collections managed by the controller servlet into collection classes (ActionForwards and ActionMappings), and publish these collections as servlet context attributes so that custom tags may access them.
Create a new <struts:forward name="xxx"/> tag that will look up the
specified logical name in the global forwards collection published above,
and perform a forward or redirect to the corresponding actual path.  The
remainder of the current page will be skipped.

Create a new <struts:include name="xxx"/> tag that will look up the
specified logical name in the global forwards collection published above,
and perform an include of the contents of the corresponding actual path at
this point in the output of the current page.  Unfortunately, this makes
the class name ActionForward something of a misnomer :-(, but it is useful
in decoupling logical and physical page names even within the View
components.

Fix an implementation bug in the <struts:redirect> tag that erroneously
returned SKIP_PAGE from the doStartTag() method.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47713 13f79535-47bb-0310-9956-ffa450edef68