Add two new custom tags to the Struts library:
<struts:options1> is a version of <struts:options> that is compatible with
JDK 1.1, and can enumerate over an array, a Dictionary (including a
Hashtable), an Enumeration, or a Vector.

<struts:multibox> is a tag that creates checkboxes, based on Ralph
Schaer's proposed modifications to the existing <struts:checkbox> tag.  I
added this functionality as a new tag instead, to allow the older tag to
remain consistent with the other input field tags in how it handled the
optional value attribute.

The basic idea of <struts:multibox> is that you have an array-valued
property (the underlying data type can be String or any Java primitive),
and the values present in the array define which checkboxes, out of a set
of checkboxes with the same property name, should be marked.  For example,
assume you have a property named "os", and a property getter named getOs()
that returns "int[]".  You might use it like this in your JSP page:

<struts:multibox property="os" value="1">Win98</struts:multibox>
<struts:multibox property="os" value="2">Win2000</struts:multibox>
<struts:multibox property="os" value="3">Linux</struts:multibox>

Note that <struts:multibox> will also work with an existing boolean bean
property, and provides one approach to getting around the "sticky
checkbox" problem.  For example, if your bean has a boolean property named
"enabled", the following command generates a checkbox that is checked if
the property is true, and unchecked if the property is false:

<struts:multibox property="enabled" value="true">Enabled</struts:multibox>


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47726 13f79535-47bb-0310-9956-ffa450edef68