Implement a series of interconnected changes:
ActionError / ActionErrors
--------------------------
A new representation of error messages returned by validation methods that
allows you to associate particular errors with particular field names, and
stores up to for parameter replacement strings along with the message key,
so you can use {0} type replacements.

Action
------
Use the new ActionErrors object instead of ErrorMessages.

ActionMapping
-------------
Update the toString() representation.

ActionServlet
-------------
Separate the functionality to populate ActionForm properties and the
functionality to call the validation method into different protected
methods that can be overridden separately if desired.

ActionForm / ValidatingActionForm
---------------------------------
Convert ActionForm from an interface to a base class, which includes
default implementations of a validate() method.  The validate() method
also takes the current mapping and request as arguments to provide more
useful contextual information.

NOTE:  DUE TO ONGOING DISCUSSIONS ON STRUTS-DEV, IT IS PREMATURE TO
ASSUME THAT THIS CHANGE IS PERMANENT.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47819 13f79535-47bb-0310-9956-ffa450edef68