Add "event and listener infrastructure for Struts" to the 1.1 TODO list. This is going to be dramatically affected by other changes contemplated in 1.1, and it looks like we'll be able to add it without a tremendous amount of backwards compatibility pain anyway -- so I'd rather not hold up 1.0 while we figure out exactly what events are "interesting" and all the other issues related to including this now.
Add a BeanUtils method to clone a JavaBean that follows the normal
conventions, even if the class itself does not implement Cloneable so that
the clone() method works.  This is based on a patch submitted by Taylor
Cowan <taylor@bondisoftware.com>, but takes advantage of the
PropertyUtils.copyProperties() method that was not available when the
patch was submitted.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47957 13f79535-47bb-0310-9956-ffa450edef68