Refactor BeanUtils by splitting out the conversion methods to ConvertUtils and anything related to processing servlet requests into RequestUtils.  As a result, BeanUtils is now only dependent on JDK classes, making it more amenable for use in stand alone programs.
Also, changed the BeanUtils.populate() method signature to take a Map
instead of a Hashtable for the properties to be set.  In Java2, the
Hashtable class implements the Map interface, so this will not cause
problems for users who actually use a Hashtable.

TODO: Need to validate that I didn't break any functionality, and then
re-introduce nessted/indexed setters in the new arrangement.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@47986 13f79535-47bb-0310-9956-ffa450edef68