Fixes for using nested and indexed properties in HTML forms!  You can now use simple/nested/indexed expressions like this:
    <html:text property="mailingAddress.street"/>

and the appropriate processing will be applied to get the street address
(for initial display) and set the street address (when the form is
processed).

Two pages exist in the "struts-test" unit test webapp - each of them
should remember any changes you make from one screen to the next, because
they are captured automatically in an ActionForm.

TODO:  Verify that I did not break file uploads while making these
changes.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@48006 13f79535-47bb-0310-9956-ffa450edef68