Update the release notes for the next beta to include a list of the files that are included in the binary distribution, and what to do with them.
In addition, I'm starting a precedent to include other useful library
files in the "lib" directory.  The first one is the JDBC 2.0 Optional
Package APIs (javax.sql), which you need to include in your webapp if it
is not provided by your servlet container.

Other libraries that are used to create Struts components in the future
will also be included (assuming that redistribution rights allow this)
to increase the usefulness of the Struts distribution.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@48306 13f79535-47bb-0310-9956-ffa450edef68