Fix for Bugzilla #2233 (single quotes in application resources do not show up even though they are properly escaped with "\").
The underlying problem is that the messages strings in your application
resources files are processed twice -- once when loaded by Property.load()
and once when the message is formatted with MessageFormat.format().  The
first call converts occurrences of "\'" to "'" in the format string; but
MessageFormat has an undocumented requirement that single quotes be
escaped by an extra single quote (not by a backslash).  The patch escapes
the format string as expected by MessageFormat so that this actually
happens.

This affects the lookup and display of message strings in all of the
following tags:
  <html:errors>
  <html:image>
  <html:img>
  <html:messages>
  <html:option

PR: Bugzilla #2233
Submitted by:	rmatteodg@infinito.it


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@48526 13f79535-47bb-0310-9956-ffa450edef68