Fixed a bug in the toHex() method which caused leading zeros to be omitted from the string representation.
Note that I do not plan on incorporating this fix into Struts 1.0.1. This
method is currently used only to generate tokens which will be exactly
matched. As such, accuracy is less important than uniqueness. The lack of
accuracy should not constitute a showstopper for Struts 1.0.1.

PR: 5458
Submitted by: Marcelo Collao Huper


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@48761 13f79535-47bb-0310-9956-ffa450edef68