Bring the test harness for Struts 1.1 up to date with respect to changes in both the Struts classes and changes required by switching to Cactus 1.3 (which is now required).
All test execution is now in build-tests.xml, but there are convenient
execution triggers in build.xml itself:

  test.junit             Execute non-Cactus JUnit tests
  test.tomcat.32         Execute Cactus tests on Tomcat 3.2
  test.tomcat.40         Execute Cactus tests on Tomcat 4.0
  test.tomcat.41         Execute Cactus tests on Tomcat 4.1
  test tomcat.all        Execute Cactus tests on all Tomcats for which
                         you have a "tomcat.home.XX" property defined

NOTE:  You will need to define a couple of new properties to execute tests:

  cactus.home           Pathname to install directory for Cactus 1.3
  aspectjrt.jar         AspectJ runtime Jar (usually you can set this to
                        "${cactus.home}/lib/aspectjrt.jar")
  cactus.jar            Cactus JAR file
  cactus.ant.jar        Cactus Ant JAR file
  log4j.jar             Log4J JAR file

In addition, you'll need to follow the Cactus 1.3 installation rules about
putting the necessary files into your $ANT_HOME/lib directory, or use the
prepackaged Cactus Ant download that has what you need.

NOTE:  I don't have Tomcat 3.2 installed any more, so I don't know for
sure if the tests will really pass on it.  They pass on 4.0 and 4.1
(head of each CVS branch).

NOTE:  On 4.0 and 4.1, Jasper complains a lot about "Stream closed"
problems, but it doesn't seem to cause the tests to fail.  To get this
output out of the way, I've put a <Logger> in the server.xml file for
these two versions, that writes to a file in the
"target/tests/server/tomcatXX" subdirectory.

NOTE:  Cactus logging is configured to write to "cactus_client.log" and
"cactus_server.log" in your top level source directory.

Now it's time to write a whole bunch of unit tests ...


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@49115 13f79535-47bb-0310-9956-ffa450edef68