Support pluggable patterns for mapping the path of an ActionForward (if it is not marked context relative), or the path specified by the "page" attribute of many tags, so that you can gracefully deal with things like migrating your JSP pages under /WEB-INF.
The syntax for the "forwardPattern" and "pagePattern" patterns is simple:
  $C is replaced by the context path of this webapp
  $A is replaced by the prefix for this subapp
  $P is replaced by the path of the forward (for forwardPattern)
     or the specified page value (for pagePattern)
  $$ is replaced by a literal $
  all other characters are passed through unchanged.

The default pattern in each case is "$C$A$P", which is consistent with the
previous behavior (and optimized for performance in the code).  To put your
pages under "/WEB-INF/pages" instead of the context root, specify a pattern
like "$C/WEB-INF/pages$A$P" instead.

Added unit tests to validate the new behavior (including consistent
behavior of the previous defaults).

FIXME - Update ImageTag and ImgTag to use RequestUtils.computeURL() so that
they will pick up changes like this automatically.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@49246 13f79535-47bb-0310-9956-ffa450edef68