Add an optional "size" attribute on <form-property>, so that you can ask for an array of a specific size without having to specify initial values in the "initial" attribute.  For example,
  <form-property name="foo" type="java.lang.String[]" size="10"/>

will cause the initialize() method of the DynaActionForm to create an
array of ten Strings (all initialized to null).  The size attribute will be
ignored if the initial attribute is present.  Updated unit tests included.

This is the non-controversial half of the recommendations in Bugzilla #14800.
I'm still reviewing the advisability of changing the way that initial arrays
get cloned or not.

PR: Bugzilla #14800
Submitted by:	James Turner <turner at blackbear.com>


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@50011 13f79535-47bb-0310-9956-ffa450edef68