The beginnings of a standard PlugIn that attempts to verify the configuration information encapsulated in the ModuleConfig for the current module.  It logs any error reports it finds via ServletContext.log(), and optionally (but does by default) throws an initialization exception if any error was actually found.
The standard struts-example webapp has been modified to register this PlugIn.

Known Limitation:  Checking things here will not catch some kinds of errors
like defining two <forward> elements with the same "name" value.  We'll need
some additional help (perhaps in the addFoo() methods of ModuleConfig) to
catch that kind of thing.

This PlugIn is at least partially in response to Bugzilla #14749, which
reports something that really is pilot error (incorrect value in the
configuration file), but is something that is hard for a newbie to track
down.  We need to reduce the number of errors that are hard to trace.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@50123 13f79535-47bb-0310-9956-ffa450edef68