Updates to announce Struts 1.1 Release Candidate 1.
IMPORTANT: Please do not update the web site with this or later content
until you see the official announcement message for the Struts 1.1 RC1
release. I'm updating these pages now so that they are included in the
RC1 struts-documentation.war web app, but I have not finished testing
the release itself.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@50270 13f79535-47bb-0310-9956-ffa450edef68