Tweaks to ComposableRequestProcessor and CatalogConfiguratorPlugIn so that the canonical Struts example application works successfully with it (at least to a first order of approximation -- not every detail has been checked).
The src/conf/chain-config.xml resource now documents the correspondence
between the standard methods of RequestProcessor versus the Command
implementations that perform those tasks in this environment.  This will
also serve as a checklist of what isn't implemeneted yet, so please keep
it up to date if you add in some of the missing features.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@51038 13f79535-47bb-0310-9956-ffa450edef68