Add the beginnings of an "application controller" filter that, at the moment, just registers the ImplPhaseListener with JSF to provide the prepare() and destroy() calls part of the ViewController contract.  This is enough to start writing sample apps that only use the view controller functionality of the shale proposal, so that we can (a) make sure what's here works, and (b) examine the example source code for simplicity and usability versus the traditional options.
Side note -- one of the benefits of Servlet 2.4 is that we're guaranteed
that servlet context listeners have all been called before the filter
initialization takes place.  This is important because we're assuming that
JSF has been initialized before the init() method of this filter is called.
That's guaranteed to be true for either the JSF RI or MyFaces (both of which
use a servlet context listener for their setup) in a Servlet 2.4 environment.
In a Servlet 2.3 environment, the order of initialization calls is not
defined, so it's a hope, rather than a certainty, that JSF will have been
initialized already.


git-svn-id: https://svn.apache.org/repos/asf/struts/trunk@55676 13f79535-47bb-0310-9956-ffa450edef68