Add a default value for struts.el.home. I'm not completely convinced that this is a good idea, because it is based on two assumptions:
1) The checked out path matches the repo path.
2) The trunk is being built.

The second of these could potentially cause problems for someone trying to build a branch or tag, forgetting to set this property to also specify the branch or tag.

However, I figured that overall, having the default might cause fewer breakages for people - at least people who retain the repo structure on their local disk (even if I'm not one of that group ;).

git-svn-id: https://svn.apache.org/repos/asf/struts/core/trunk@65585 13f79535-47bb-0310-9956-ffa450edef68