Add the ability to FormBeanConfig to create an ActionForm based on an ActionContext instead of a Servlet, hiding a dependency on the Servlet API.
This enables the elimination of o.a.s.chain.commands.AbstractCreateForm and it's subclass in the o.a.s.c.c.servlet package.  These are deprecated, but will be removed before the first Struts 1.3 release, but are preserved temporarily to smooth the transition for users who have customized the default chain-config.xml file.  (In that file, "org.apache.struts.chain.commands.servlet.CreateActionForm" should be changed to "org.apache.struts.chain.commands.CreateActionForm".

Change chain-config.xml to use new version of CreateActionForm.

Add convenience method to ActionContextBase to findOrCreateForm given a form name and scope name, since it contains all the necessary referents to achieve that.

Add logger property to ActionContextBase, with default initialization strategy.

Create generic command which can use configuration parameters to lookup or create an ActionForm and place it in the ActionContext under a specified key, so that a subsequent command can prepopulate fields.

Centralize logic to determine whether a given ActionForm instance can be "re-used" for a given FormBeanConfig as an instance method of FormBeanConfig.  Adjust RequestUtils to use newly centralized logic.




git-svn-id: https://svn.apache.org/repos/asf/struts/core/trunk@161497 13f79535-47bb-0310-9956-ffa450edef68