Extend arbitrary property maps from ActionConfig to nearly all config objects by introducing BaseConfig class.  Add some explanation to the DTD about it. For Bugzilla enhancement request 34691; thanks to Riyaz Mansoor for the basic patch and the motivation.
Note that this does not apply to the PlugInConfig object because 
(a) that object is not very public to developers, 
(b) developers must always create a concrete implementation of PlugIn, and so are able to create bean properties as necessary
(c) PlugInConfig already had a "properties" object which would have conflicted with the one in BaseConfig.  

If people are pained by this small inconsistency or clamor for the feature to be extended, we can revisit.



git-svn-id: https://svn.apache.org/repos/asf/struts/core/trunk@169898 13f79535-47bb-0310-9956-ffa450edef68