Modify ExceptionHandler to take into account the case when the HttpServletResponse is "committed". This happens often when Tiles is being used, and results in the dreaded "blank screen."
Henceforth, by default, the ExceptionHandler tests response.isCommitted() and when true, in the normal case, simply
"includes" the configured view path, rather than forwarding to it. 

For users who want to use an alternate view path in this case (perhaps a simpler one which fits better into
a partially rendered tileset), it is possible to set an alternate path for use only in this case, and for users
who don't want anything done when the response is committed, that can also be configured.  See the JavaDoc for details.

git-svn-id: https://svn.apache.org/repos/asf/struts/core/trunk@209532 13f79535-47bb-0310-9956-ffa450edef68